feat(): support for setting challenges on subdomains, closes #6
certbot takes one or more -d arguments to specify the domain(s) for
which a certificate is requested. So far, the hook script assumed that
only one such argument (-d $DEDYN_NAME) was passed to certbot.

With this commit, the hook script determines whether the challenge that
is going to be set is for a subdomain of the zone, and if so, it will
add an infix to the subname field when setting the challenge RRset. The
subname field then reads: _acme-challenge$infix

As certbot calls the hook script once per -d argument, this behavior
adds support for subdomains. I have not tested whether this works for
wildcard subdomains; however, I am not aware of wildcard limitations
in this script.

This commit also switches to bulk REST requests. This comes with the
advantage that one request per challenge suffices, regardless of
whether the RRset existed before.