Always wait 120s for challenge propagation
We can't efficiently test if the challenge has already been
propagated everywhere, unless we iterate over all frontend
servers. This can cause a situation where the script assumes
the challenge is published, because the nearest frontend
already applied the changes, but LE's nearest frontend did
not yet publish it. In this case, certificate issuance will
fail.